PHOENIX-6327 Consolidate Junit test categories
