+++ phoenix-core/src/it/java/org/apache/phoenix/end2end/ParallelStatsDisabledTest.java
@@ -32,8 +32,8 @@ import org.apache.phoenix.query.BaseTest;
  * these tests use random table names generated by {@link BaseTest#generateRandomString}.
  * 
  * Tests using a mini cluster need to be classified either 
- * as {@link ParallelStatsDisabledTest} or {@link ClientManagedTimeTest}
- * or {@link NeedsOwnMiniClusterTest} or {@link HBaseManagedTimeTest} otherwise they won't be run
+ * as {@link ParallelStatsDisabledTest} or {@link ParallelStatsEnabledTest}
+ * or {@link NeedsOwnMiniClusterTest} otherwise they won't be run
  * when one runs mvn verify or mvn install.
  * 
  * @since 4.1
