+++ phoenix-core/src/test/java/org/apache/phoenix/query/BaseTest.java
@@ -135,7 +135,6 @@ import org.apache.hadoop.hbase.regionserver.HRegionServer;
 import org.apache.hadoop.hbase.regionserver.RSRpcServices;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.apache.phoenix.SystemExitRule;
-import org.apache.phoenix.end2end.BaseHBaseManagedTimeIT;
 import org.apache.phoenix.exception.SQLExceptionCode;
 import org.apache.phoenix.exception.SQLExceptionInfo;
 import org.apache.phoenix.jdbc.PhoenixConnection;
@@ -171,13 +170,19 @@ import org.apache.phoenix.thirdparty.com.google.common.util.concurrent.ThreadFac
  * Base class that contains all the methods needed by
  * client-time and hbase-time managed tests.
  * 
- * For tests needing connectivity to a cluster, please use
- * {@link BaseHBaseManagedTimeIT}.
+ *  * Tests using a mini cluster need to be classified either 
+ * as {@link ParallelStatsDisabledTest} or {@link ParallelStatsEnabledTest}
+ * or {@link NeedsOwnMiniClusterTest} otherwise they won't be run
+ * when one runs mvn verify or mvn install.
  * 
- * In the rare case when a test can't share the same mini cluster as the 
- * ones used by {@link BaseHBaseManagedTimeIT},
- * one could extend this class and spin up your own mini cluster. Please 
- * make sure to shutdown the mini cluster in a method annotated by @AfterClass.  
+ * For tests needing connectivity to a cluster, please use
+ * {@link ParallelStatsDisabledIt} or {@link ParallelStatsEnabledIt}.
+ * 
+ * In the case when a test can't share the same mini cluster as the 
+ * ones used by {@link ParallelStatsDisabledIt} or {@link ParallelStatsEnabledIt},
+ * one could extend this class and spin up your own mini cluster. Please
+ * make sure to annotate such clesses with {@link NeedsOwnMiniClusterTest} and
+ * shutdown the mini cluster in a method annotated by @AfterClass.
  *
  */
